Домашнє завдання №3 Node
