+++ middlewares/index.js
@@ -1,0 +1,6 @@
+const ctrlWrapper = require("./ctrlWrapper");
+const validation = require("./validation");
+module.exports = {
+  ctrlWrapper,
+  validation,
+};
