+++ schema/index.js
@@ -1,0 +1,4 @@
+const contactSchema = require("./contacts");
+module.exports = {
+  contactSchema,
+};
